terbaru

+++ content/tambah-instructor-major.php
@@ -10,40 +10,98 @@ if (isset($_GET['delete'])) {
 
     $queryDelete = mysqli_query($config, "DELETE FROM instructor_majors WHERE id='$id'");
     if ($queryDelete) {
-        header("location:?page=tambah-instructor-major&id=" . $id . "&hapus=berhasil");
+        header("location:?page=tambah-instructor-major&id=" . $id_instructor . "&hapus=berhasil");
     } else {
-        header("location:?page=tambah-instructor-major&id=" . $id . "&tambah=gagal");
+        header("location:?page=tambah-instructor-major&id=" . $id_instructor . "&tambah=gagal");
     }
 }
 
-$id = isset($_GET['id']) ? $_GET['id'] : '';
+$id_instructor = isset($_GET['id']) ? $_GET['id'] : '';
+$edit = isset($_GET['edit']) ? $_GET['edit'] : '';
 // print_r($id_user);
 if (isset($_POST['id_major'])) {
     //ada tidak parameter bernama edit, kalo ada jalankan perintah edit/update, kalo tidak ada tambah data baru/insert
     $id_major = $_POST['id_major'];
+    if (isset($_GET['edit'])) {
+        $update = mysqli_query($config, "UPDATE instructor_majors SET id_major = '$id_major' WHERE id='$edit'");
+        header("location:?page=tambah-instructor-major&id=" . $id_instructor . "&ubah=berhasil");
+    } else {
 
-    $insert = mysqli_query($config, "INSERT INTO instructor_majors (id_major, id_instructor) 
- VALUES ('$id_major', '$id')");
-    header("location:?page=tambah-instructor-major&id=" . $id . "&tambah=berhasil");
+        $insert = mysqli_query($config, "INSERT INTO instructor_majors (id_major, id_instructor) 
+     VALUES ('$id_major', '$id_instructor')");
+        header("location:?page=tambah-instructor-major&id=" . $id_instructor . "&tambah=berhasil");
+    }
 }
 
 $queryMajor = mysqli_query($config, "SELECT * FROM majors ORDER BY id DESC");
 $rowMajors = mysqli_fetch_all($queryMajor, MYSQLI_ASSOC);
 
-$queryInstructor = mysqli_query($config, "SELECT * FROM instructors WHERE id = '$id'");
+$queryInstructor = mysqli_query($config, "SELECT * FROM instructors WHERE id = '$id_instructor'");
 $rowInstructor = mysqli_fetch_assoc($queryInstructor);
 
 $queryInstructorMajor = mysqli_query($config, "SELECT majors.name, instructor_majors.id, id_instructor FROM instructor_majors 
-LEFT JOIN majors ON majors.id = instructor_majors.id_major WHERE id_instructor = '$id' ORDER BY instructor_majors.id DESC");
+LEFT JOIN majors ON majors.id = instructor_majors.id_major WHERE id_instructor = '$id_instructor' ORDER BY instructor_majors.id DESC");
 $rowInstructorMajors = mysqli_fetch_all($queryInstructorMajor, MYSQLI_ASSOC);
 
+$queryEdit = mysqli_query($config, "SELECT * FROM instructor_majors WHERE id='$edit'");
+$rowEdit = mysqli_fetch_assoc($queryEdit);
 ?>
 
 <div class="row">
     <div class="col-sm-12">
         <div class="card">
             <div class="card-body">
-                <h5 class="card-title"> Add Instructor Major : <?php echo $rowInstructor['name'] ?></h5>
+                <h5 class="card-title"><?php echo isset($_GET['edit']) ? 'Edit' : 'Add' ?> Instructor Major : <?php echo $rowInstructor['name'] ?></h5>
+                <!-- form edit -->
+                <?php if (isset($_GET['edit'])): ?>
+                    <form action="" method="post">
+                        <div class="mb-3">
+                            <label for="">Major Name</label>
+                            <select name="id_major" id="" class="form-select" aria-label="Default select example">
+                                <option value="">Select One</option>
+                                <?php foreach ($rowMajors as $rowMajor): ?>
+                                    <option <?php echo $rowMajor['id'] == $rowEdit['id_major'] ? 'selected' : '' ?> value=" <?php echo $rowMajor['id'] ?>" <?php echo $rowMajor['name'] ?>></option>
+                                <?php endforeach ?>
+                            </select>
+                        </div>
+                        <div class="mb-3">
+                            <button class="btn btn-primary" type="submit">Save Changes</button>
+                        </div>
+                    </form>
+                    <!-- end form edit  -->
+                <?php else: ?>
+                    <!-- Button trigger modal -->
+                    <!-- Modal -->
+                    <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                        <div class="modal-dialog">
+                            <div class="modal-content">
+                                <div class="modal-header">
+                                    <h1 class="modal-title fs-5" id="exampleModalLabel">Add New Instructor Major : </h1>
+                                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                </div>
+                                <form action="" method="post">
+                                    <div class="modal-body">
+                                        <div class="mb-3">
+                                            <label for="">Major Name</label>
+                                            <select name="id_major" id="" class="form-select" aria-label="Default select example">
+                                                <option value="">Select One</option>
+                                                <?php foreach ($rowMajors as $rowMajor): ?>
+                                                    <option value=" <?php echo $rowMajor['id'] ?>"><?php echo $rowMajor['name'] ?></option>
+                                                <?php endforeach ?>
+                                            </select>
+                                        </div>
+                                        ...
+                                    </div>
+                                    <div class="modal-footer">
+                                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                                        <button type="submit" class="btn btn-primary">Save changes</button>
+                                    </div>
+                                </form>
+                            </div>
+                        </div>
+                    </div>
+                <?php endif ?>
+                <!-- listing table -->
                 <div align="right">
                     <button type="button" class="btn btn-primary mb-3" data-bs-toggle="modal" data-bs-target="#exampleModal">
                         Add Instructor Major
@@ -54,7 +112,7 @@ $rowInstructorMajors = mysqli_fetch_all($queryInstructorMajor, MYSQLI_ASSOC);
                         <tr>
                             <th>No</th>
                             <th>Major Name</th>
-                            <th></th>
+                            <th>Action</th>
                         </tr>
                     </thead>
                     <tbody>
@@ -65,47 +123,15 @@ $rowInstructorMajors = mysqli_fetch_all($queryInstructorMajor, MYSQLI_ASSOC);
                                 <td><?php echo $no++ ?></td>
                                 <td><?php echo $rowInstructorMajor['name']; ?></td>
                                 <td>
-                                    <a onclick="return confirm('Are you sure wanna delete this data??')"
-                                        href="?page=tambah-instructor-major&delete=<?php echo $rowInstructorMajor['id'] ?>
-                                        &id_instructor=<?php echo $rowInstructorMajor['id_instructor'] ?>"
-                                        class="btn btn-danger btn-sm">Delete</a>
+                                    <a href="?page=tambah-instructor-major&id=<?php echo $rowInstructorMajor['id_instructor'] ?>&edit=<?php echo $rowInstructorMajor['id'] ?>" class="btn btn-primary btn-sm">Edit</a>
+                                    <a onclick="return confirm('Are you sure wanna delete this data??')" href="?page=tambah-instructor-major&delete=<?php echo $rowInstructorMajor['id'] ?> &id_instructor=<?php echo $rowInstructorMajor['id_instructor'] ?>" class="btn btn-danger btn-sm">Delete</a>
                                 </td>
                             </tr>
                         <?php endforeach ?>
                     </tbody>
                 </table>
-                <!-- Button trigger modal -->
 
 
-                <!-- Modal -->
-                <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-                    <div class="modal-dialog">
-                        <div class="modal-content">
-                            <div class="modal-header">
-                                <h1 class="modal-title fs-5" id="exampleModalLabel">Add New Instructor Major : </h1>
-                                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                            </div>
-                            <form action="" method="post">
-                                <div class="modal-body">
-                                    <div class="mb-3">
-                                        <label for="">Major Name</label>
-                                        <select name="id_major" id="" class="form_control">
-                                            <option value="">Select One</option>
-                                            <?php foreach ($rowMajors as $rowMajor): ?>
-                                                <option value="<?php echo $rowMajor['id'] ?>"><?php echo $rowMajor['name'] ?></option>
-                                            <?php endforeach ?>
-                                        </select>
-                                    </div>
-                                    ...
-                                </div>
-                                <div class="modal-footer">
-                                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                                    <button type="submit" class="btn btn-primary">Save changes</button>
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </div>
             </div>
         </div>
     </div>
